firewall optimizing and refactoring
Major refactor of the code to help keep concepts clear as I added in firewall optimizing. The 'offset scanner' approach means scanners of the same size can be merged together to a single scanner and small scanners that are exactly half, a third, a quarter, etc. the size of another scanner can be expanded and merged as well. With the example firewall, 43 scanners in the firewall are reduced to 5 using these techniques. From the previous build, this doubles performance. From fast_mode.py, this is ~11x faster and easily runs under one second.